Use GitHub App integration over OAuth App for authentication
This removes the need for user scopes and replaces the default dev
credentials for our applications with default credentials for our
dev GitHub App. User scopes are no longer required when using a GitHub
App, the permissions are instead based on what the user grants what
the App is asking for and what they make available.

Emails aren't guraranteed to be available to the GitHub App API so
we are optionally checking for it's value before setting it.
Additionally teams are handled differently and we will need to
do a bit of work to re-enable feature flagging on the back end.
We aren't currently using it so this shouldn't matter too much.

+++ components/builder-sessionsrv/src/config.rs
@@ -71,14 +71,14 @@ impl AppCfg for Config {
 pub struct PermissionsCfg {
     /// A GitHub Team identifier for which members will automatically have administration
     /// privileges assigned to their session
-    pub admin_team: u64,
+    pub admin_team: u32,
     /// GitHub team's whose members are granted Builder Worker abilities. These abilities
     /// include downloading the latest private key for any origin and uploading a package into
     /// any origin regardless of membership. This is essentially a user who ignores all rules.
-    pub build_worker_teams: Vec<u64>,
+    pub build_worker_teams: Vec<u32>,
     /// GitHub team's assigned to the early access group who may have access to features in Builder
     /// before a normal user.
-    pub early_access_teams: Vec<u64>,
+    pub early_access_teams: Vec<u32>,
 }
 
 #[cfg(test)]
